[tg] Refactor `get_ids`
