+++ tg/handlers.py
@@ -10,6 +10,7 @@ from tg import (
     code_runner,
     stats,
     others,
+    utils,
 )
 
 _logger = logging.getLogger(__name__)
@@ -30,7 +31,7 @@ HANDLERS = [
     ),
     # commands
     handlers.MessageHandler(
-        get_ids.choose_lang,
+        others.choose_lang,
         filters.private
         & ~filters.tg_business
         & tg_filters.start_command(command="lang")
@@ -54,7 +55,7 @@ HANDLERS = [
         & tg_filters.create_user(),
     ),
     handlers.MessageHandler(
-        get_ids.added_to_group,
+        others.added_to_group,
         filters.private
         & ~filters.tg_business
         & tg_filters.start_command(command="add")
@@ -77,7 +78,7 @@ HANDLERS = [
         & tg_filters.create_user(),
     ),
     handlers.MessageHandler(
-        get_ids.send_about,
+        others.send_about,
         filters.private
         & ~filters.tg_business
         & tg_filters.start_command(command="about")
@@ -93,7 +94,7 @@ HANDLERS = [
         & tg_filters.create_user(),
     ),
     handlers.MessageHandler(
-        get_ids.send_privacy_policy,
+        others.send_privacy_policy,
         filters.private
         & ~filters.tg_business
         & tg_filters.start_command("privacy")
@@ -101,7 +102,7 @@ HANDLERS = [
         & tg_filters.create_user(),
     ),
     handlers.MessageHandler(
-        get_ids.send_link_to_chat_by_id,
+        utils.send_link_to_chat_by_id,
         filters.private
         & ~filters.tg_business
         & tg_filters.start_command("link")
@@ -183,7 +184,7 @@ HANDLERS = [
         & tg_filters.create_user(),
     ),
     handlers.ChatMemberUpdatedHandler(
-        get_ids.on_remove_permission,
+        others.on_remove_permission,
     ),
     # business
     handlers.MessageHandler(
@@ -203,7 +204,7 @@ HANDLERS = [
         & tg_filters.create_user(),
     ),
     handlers.BusinessBotConnectionHandler(
-        get_ids.handle_business_connection,
+        others.handle_business_connection,
         tg_filters.create_user(),
     ),
     # welcome
@@ -223,7 +224,7 @@ HANDLERS = [
         & tg_filters.create_user(),
     ),
     handlers.CallbackQueryHandler(
-        get_ids.get_lang,
+        others.get_lang,
         filters.create(lambda _, __, cbd: cbd.data.startswith("lang"))
         & tg_filters.is_user_spamming()
         & tg_filters.create_user(),
